+++ src/scripts/pages/home.js
@@ -1,4 +1,4 @@
-import photographerFactory from '../factories/photographer.js';
+import photographerFactory from '../factories/photographerFactory.js';
 
 // Fetch photographers data
 export async function getPhotographers() {