+++ src/scripts/pages/photographer.js
@@ -1,57 +1,4 @@
-/*
-! Import not working, needs to be fixed
- import photographerFactory from './src/scripts/factories/photographer.js'; */
-
-// ! Beginning of temporary code. Need to use "import" instead
-/* 
-! Code used for "Method 1 only "
-function photographerFactory(data) {
-    const { name, city, tagline, price, id, portrait } = data;
-
-    const picture = `assets/photographers/${portrait}`;
-
-    function getUserCardDOM() {
-        const article = document.createElement('article');
-
-        // Redirection link
-        const photographerLink = document.createElement('a');
-        photographerLink.setAttribute(
-            'href',
-            `./pages/photographer-pages/photographer.html?id=${id}`
-        );
-        article.appendChild(photographerLink);
-
-        // Profile picture
-        const img = document.createElement('img');
-        img.setAttribute('src', picture);
-        photographerLink.appendChild(img);
-
-        // Photographer name
-        const h2 = document.createElement('h2');
-        h2.textContent = name;
-        photographerLink.appendChild(h2);
-
-        // Location
-        const location = document.createElement('p');
-        location.textContent = city;
-        article.appendChild(location);
-
-        // Motto
-        const motto = document.createElement('p');
-        motto.textContent = tagline;
-        article.appendChild(motto);
-
-        // Pricing
-        const pricing = document.createElement('p');
-        pricing.textContent = price;
-        article.appendChild(pricing);
-
-        return article;
-    }
-    return { name, picture, city, tagline, id, price, getUserCardDOM };
-} */
-
-// ! End of temporary code.
+import photographerFactory from '../factories/photographerFactory.js';
 
 // Retrieve the id of the photographer from the URL and convert in number to find It with ".find()" with strict equality
 const params = new URL(document.location).searchParams;
@@ -92,26 +39,12 @@ export async function getMedia() {
     }
 }
 
-/*  
-? Display Method 1
 // Display photographer profile data on page
 async function displayData(photographerProfile) {
     const photographHeader = document.querySelector('.photographer-info');
     const photographerModel = photographerFactory(photographerProfile);
-    const userCardDOM = photographerModel.getUserCardDOM();
+    const userCardDOM = photographerModel.getProfile();
     photographHeader.appendChild(userCardDOM);
-} */
-
-// ? Display Method 2
-// Display photographer profile data on page
-export async function displayDataProfile(profile) {
-    const profileHeaderHtml = `<img src="/assets/photographers/${profile.portrait}">
-    <h1>${profile.name}</h1>
-    <h2>${profile.city} ${profile.country}</h2>
-    <p>${profile.tagline}</p>`;
-
-    const photographHeader = document.querySelector('.photographer-info');
-    photographHeader.innerHTML = profileHeaderHtml;
 }
 
 // Display photographer medias on page
@@ -127,6 +60,6 @@ export async function displayMedia(media) {
 export default async function initPhotographer() {
     const profiles = await getProfile();
     const medias = await getMedia();
-    displayDataProfile(profiles);
+    displayData(profiles);
     displayMedia(medias);
 }